refactor: contextprovider 컴포넌트를 사용하도록 리팩토링
