+++ src/components/CardHolderNameInput.tsx
@@ -1,11 +1,11 @@
 import React, { useCallback, useContext, useState } from "react";
-import CardInfoContext from "context/CardInfoContext";
 
 import { CARD_INFO_RULES } from "utils/constants";
 
 import Input from "components/UIComponents/Input/Input";
 import InputField from "components/UIComponents/InputField/InputField";
 import WordCounter from "components/UIComponents/WordCounter/WordCounter";
+import { CardInfoContext } from "context/CardInfoContextProvider";
 
 export default function CardHolderNameInput() {
   const [isInvalid, setInvalid] = useState(false);
